change Reactive Warhammer to Vigorous Warhammer
