+++ source/gear/_gearDictionary.js
@@ -213,9 +213,9 @@ for (const file of [
 	"warcry-slowing.js",
 	"warcry-tormenting.js",
 	"warhammer-base.js",
-	"warhammer-reactive.js",
 	"warhammer-slowing.js",
 	"warhammer-unstoppable.js",
+	"warhammer-vigorous.js",
 	"wolfring-base.js",
 	"wolfring-surpassing.js",
 	"wolfring-swift.js",
