--- source/gear/warhammer-reactive.js
@@ -1,39 +1,0 @@
-const { GearTemplate, Move } = require('../classes');
-const { ELEMENT_MATCH_STAGGER_FOE } = require('../constants.js');
-const { dealDamage, changeStagger } = require('../util/combatantUtil.js');
-
-module.exports = new GearTemplate("Reactive Warhammer",
-	[
-		["use", "Strike a foe for <@{damage} + @{bonus} if foe is stunned + @{bonus2} if after foe> @{element} damage"],
-		["Critical💥", "Damage x@{critMultiplier}"]
-	],
-	"Weapon",
-	"Earth",
-	350,
-	([target], user, adventure) => {
-		const { element, damage, bonus, critMultiplier, bonus2 } = module.exports;
-		let pendingDamage = user.getPower() + damage;
-		if (target.isStunned) {
-			pendingDamage += bonus;
-		}
-		const userMove = adventure.room.findCombatantMove({ index: adventure.getCombatantIndex(user), team: user.team });
-		const targetMove = adventure.room.findCombatantMove({ index: adventure.getCombatantIndex(target), team: target.team });
-
-		if (Move.compareMoveSpeed(userMove, targetMove) > 0) {
-			pendingDamage += bonus2;
-		}
-
-		if (user.element === element) {
-			changeStagger([target], user, ELEMENT_MATCH_STAGGER_FOE);
-		}
-		if (user.crit) {
-			pendingDamage *= critMultiplier;
-		}
-		return dealDamage([target], user, pendingDamage, false, element, adventure);
-	}
-).setTargetingTags({ type: "single", team: "foe" })
-	.setSidegrades("Slowing Warhammer", "Unstoppable Warhammer")
-	.setCooldown(1)
-	.setDamage(40)
-	.setBonus(75) // Awesome damage
-	.setBonus2(75); // Reactive damage
