+++ source/labyrinths/debugdungeon.js
@@ -29,7 +29,7 @@ module.exports = new LabyrinthTemplate("Debug Dungeon",
 			"Devoted Carrot",
 			"Lucky Carrot",
 			"Reinforced Carrot",
-			"Reactive Warhammer",
+			"Vigorous Warhammer",
 			"Reactive Spear",
 			"Stick",
 			"Staggering Stick",
